backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
numpy==1.24.3
pandas==2.0.3
plotly==5.17.0
python-dotenv==1.0.0
SQLAlchemy==2.0.21
openpyxl==3.1.2
google-generativeai==0.3.2
typing-extensions==4.8.0
Werkzeug==2.3.7